[FEATURE/VG-29] 파일 이벤트 추가, 에셋 브라우저 편의성 개선

+++ Source/GA6thFinal_Framework/GameEngine/Source/Engine/FileSystem/Extra/FileContext.cpp
@@ -58,8 +58,11 @@ namespace File
     {
         if (false == _path.empty())
         {
-            // 차후 동작은 이벤트로 처리하므로 파일만 변경
-            return File::RemoveFile(_path);
+            if (true == fs::exists(_path))
+            {
+                // 차후 동작은 이벤트로 처리하므로 파일만 변경
+                return File::RemoveFile(_path);
+            }
         }
         return false;
     }
@@ -70,11 +73,20 @@ namespace File
         {
             if (newPath != _path)
             {
-                // 차후 동작은 이벤트로 처리하므로 파일만 변경
-                return MoveFileExW(_path.wstring().c_str(), newPath.wstring().c_str(), MOVEFILE_REPLACE_EXISTING);
+                File::Path from = _path;
+                File::Path to = GenerateUniquePath(newPath);
+                fs::rename(from, to);
+                return true;
             }
         }
         return false;
+    }
+
+    std::weak_ptr<FolderContext> Context::GetParentContext()
+    {
+        File::Path parentPath = _path.parent_path();
+        auto       context    = UmFileSystem.GetContext<FolderContext>(parentPath);
+        return context;
     }
 
     FileContext::FileContext(const File::Path& path) 
@@ -86,7 +98,11 @@ namespace File
     {
     }
 
-    void FileContext::OnFileAdded(const Path& path) 
+    void FileContext::OnFileRegistered(const File::Path& path) 
+    {
+    }
+
+    void FileContext::OnFileUnregistered(const File::Path& path) 
     {
     }
 
@@ -102,43 +118,17 @@ namespace File
     void FileContext::OnFileRenamed(const Path& oldPath, const Path& newPath) 
     {
         _path = newPath;
-        _name = newPath.filename().string();
+        _name = _path.filename().string();
         
         LoadMeta();
-
-        //File::Path oldParentPath    = oldPath.parent_path().generic_string();
-        //File::Path newParentPath    = newPath.parent_path().generic_string();
-        //auto       oldParentContext = UmFileSystem.GetContext<File::FolderContext>(oldParentPath);
-        //auto       newParentContext = UmFileSystem.GetContext<File::FolderContext>(newParentPath);
-        //
-        //if (false == oldParentContext.expired() && false == newParentContext.expired())
-        //{
-        //    auto spOldContext = oldParentContext.lock();
-        //    auto spNewContext = newParentContext.lock();
-        //    spOldContext->_contextTable.erase(oldPath.filename());
-        //    spNewContext->_contextTable[newPath.filename()] = UmFileSystem.GetContext(GetPath());
-        //}
     }
 
     void FileContext::OnFileMoved(const Path& oldPath, const Path& newPath) 
     {
         _path = newPath;
-        _name = newPath.filename().string();
+        _name = _path.filename().string();
 
         LoadMeta();
-
-        //File::Path oldParentPath    = oldPath.parent_path().generic_string();
-        //File::Path newParentPath    = newPath.parent_path().generic_string();
-        //auto       oldParentContext = UmFileSystem.GetContext<File::FolderContext>(oldParentPath);
-        //auto       newParentContext = UmFileSystem.GetContext<File::FolderContext>(newParentPath);
-        //
-        //if (false == oldParentContext.expired() && false == newParentContext.expired())
-        //{
-        //    auto spOldContext = oldParentContext.lock();
-        //    auto spNewContext = newParentContext.lock();
-        //    spOldContext->_contextTable.erase(oldPath.filename());
-        //    spNewContext->_contextTable[newPath.filename()] = UmFileSystem.GetContext(GetPath());
-        //}
     }
 
     FolderContext::FolderContext(const File::Path& path) 
@@ -152,7 +142,33 @@ namespace File
     
     }
 
-    void FolderContext::OnFileAdded(const Path& path) 
+    void FolderContext::MoveContext(std::weak_ptr<Context> context) 
+    {
+        //if (false == context.expired())
+        //{
+        //    auto spContext = context.lock();
+        //    auto parentContext = spContext->GetParentContext();
+        //    if (false == parentContext.expired())
+        //    {
+        //        auto spParentContext = parentContext.lock();
+        //        
+        //        spParentContext->_contextTable.find
+        //        File::Path name = spContext->GetPath().filename();
+        //        _contextTable[name] = context;
+        //    }
+        //
+        //    const Path& oldPath                 = spContext->GetPath();
+        //    const Path  newPath                 = _path / spContext->GetPath().filename();
+        //    _contextTable[spContext->GetName()] = context;
+        //    spContext->Move(newPath);
+        //}
+    }
+
+    void FolderContext::OnFileRegistered(const File::Path& path)
+    {
+    }
+
+    void FolderContext::OnFileUnregistered(const File::Path& path) 
     {
     }
 
@@ -168,7 +184,7 @@ namespace File
         {
             if (false == wpContext.expired())
             {
-                UmFileSystem.RemovedFile(wpContext.lock()->GetPath());
+                UmFileSystem.ProcessRemovedFile(wpContext.lock()->GetPath());
             }
         }
     }
@@ -176,9 +192,11 @@ namespace File
     void FolderContext::OnFileRenamed(const Path& oldPath, const Path& newPath)
     {
         _path = newPath;
-        _name = newPath.filename().string();
+        _name = _path.filename().string();
 
         LoadMeta();
+
+        std::vector<std::pair<FString, FString>>  table;
 
         for (auto& [name, wpContext] : _contextTable)
         {
@@ -186,18 +204,24 @@ namespace File
             {
                 auto        spContext      = wpContext.lock();
                 const Path& oldContextPath = spContext->GetPath();
-                const Path& newContextPath = newPath / spContext->GetPath().filename();
-                UmFileSystem.MovedFile(oldContextPath.generic_string(), newContextPath.generic_string());
+                const Path newContextPath = _path / spContext->GetPath().filename();
+                table.push_back(std::make_pair(oldContextPath.generic_string(), newContextPath.generic_string()));
             }
+        }
+        for (auto& [oldPath, newPath] : table)
+        {
+            UmFileSystem.ProcessMovedFile(oldPath, newPath);
         }
     }
 
     void FolderContext::OnFileMoved(const Path& oldPath, const Path& newPath) 
     {
         _path = newPath;
-        _name = newPath.filename().string();
+        _name = _path.filename().string();
 
         LoadMeta();
+
+        std::vector<std::pair<FString, FString>> table;
 
         for (auto& [name, wpContext] : _contextTable)
         {
@@ -205,9 +229,13 @@ namespace File
             {
                 auto        spContext      = wpContext.lock();
                 const Path& oldContextPath = spContext->GetPath();
-                const Path& newContextPath = newPath / spContext->GetPath().filename();
-                UmFileSystem.MovedFile(oldContextPath, newContextPath);
+                const Path newContextPath = _path / spContext->GetPath().filename();
+                table.push_back(std::make_pair(oldContextPath.generic_string(), newContextPath.generic_string()));
             }
+        }
+        for (auto& [oldPath, newPath] : table)
+        {
+            UmFileSystem.ProcessMovedFile(oldPath, newPath);
         }
     }
